requirements.txt
asgiref==3.6.0
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.38.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.7.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
response==0.5.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
zipp==3.14.0
